🩹 Refactor: View Extension 통일화
