--- PADO/PADO/iOS/Helpers/Extensions/SettingCell+.swift
@@ -1,29 +1,0 @@
-//
-//  SettingCell+.swift
-//  PADO
-//
-//  Created by 황민채 on 1/16/24.
-//
-
-import SwiftUI
-
-// MARK: - 세팅셀 모디파이어
-extension View {
-    func settingMidCellModi() -> some View {
-        self
-            .foregroundStyle(.gray)
-            .fontWeight(.semibold)
-            .font(.system(.footnote))
-            .padding(.horizontal, UIScreen.main.bounds.width * 0.05)
-    }
-}
-
-// MARK: - 세팅셀 폰트 모디파이어
-extension View {
-    func settingFontModi() -> some View {
-        self
-            .foregroundStyle(.white)
-            .font(.system(.body))
-            .fontWeight(.semibold)
-    }
-}
